+++ CHANGELOG.md
@@ -17,6 +17,7 @@ project adheres to
   `FileContext` and `FsF ileContext` (PR #132).
 * `Scope::get` now takes an `&Name` rather than an `&str` for the
   variable name.
+* `sass::Value::Variable` now has a source position.
 
 ### Improvements